rename conv2state package to stateconverter
