+++ worker/stateconverter/converter.go
@@ -1,7 +1,7 @@
 // Copyright 2015 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package conv2state
+package stateconverter
 
 import (
 	"github.com/juju/errors"
