+++ worker/stateconverter/interfaces.go
@@ -1,7 +1,7 @@
 // Copyright 2022 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package conv2state
+package stateconverter
 
 import (
 	"github.com/juju/names/v4"
